backend/requirements.txt
flask
torch
torchvision
opencv-python
pillow
realesrgan
gfpgan